course_stats view: cast MAX(round_index) to bigint to preserve column type
CREATE OR REPLACE VIEW refuses to change a column's data type. The
earlier 20260518_course_stats_view used COUNT(*) (bigint); switching
to MAX(round_index) (integer) without a cast errors with 42P16
"cannot change data type of view column" on staging where the
previous view already exists.

+++ supabase/migrations/20260518_course_stats_from_round_index.sql
@@ -16,10 +16,14 @@
 -- CREATE OR REPLACE: the previous migration created the view; this one
 -- replaces its body. Idempotent.
 
+-- Cast to bigint so the column type matches the previous definition
+-- (COUNT(*) is bigint, MAX(integer) is integer). CREATE OR REPLACE VIEW
+-- refuses to change a column's data type, so without the cast this
+-- migration errors with 42P16 against the existing view.
 CREATE OR REPLACE VIEW course_stats AS
 SELECT
   course_code,
-  MAX(round_index) AS lego_count
+  MAX(round_index)::bigint AS lego_count
 FROM course_round_index
 GROUP BY course_code;
